lint: apply clippy

+++ src/operators/parse_operator.rs
@@ -68,7 +68,7 @@ pub fn chunk_document(document: String) -> Vec<String> {
             .join("");
         groups.push(group);
         sentences.drain(0..group_size);
-        remainder -= remainder_per_group as f32;
+        remainder -= remainder_per_group;
     }
 
     while !sentences.is_empty() {
